Refactor SE CORS under webserver. Work on package-info JavaDoc a bit
