+++ webserver/cors/pom.xml
@@ -19,24 +19,21 @@
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
     <parent>
-        <groupId>io.helidon</groupId>
-        <artifactId>helidon-project</artifactId>
+        <groupId>io.helidon.webserver</groupId>
+        <artifactId>helidon-webserver-project</artifactId>
         <version>2.0.0-SNAPSHOT</version>
     </parent>
 
-    <groupId>io.helidon.cors</groupId>
-    <artifactId>helidon-cors</artifactId>
+    <groupId>io.helidon.webserver.cors</groupId>
+    <artifactId>helidon-webserver-cors</artifactId>
     
-    <name>Helidon CORS</name>
+    <name>Helidon Webserver CORS</name>
 
     <description>
         Helidon CORS implementation
     </description>
 
     <packaging>jar</packaging>
-
-    <properties>
-    </properties>
 
     <dependencies>
         <dependency>
