+++ webserver/cors/src/test/resources/application.yaml
