+++ webserver/cors/src/test/resources/twoCORS.yaml
